requirements.txt
langchain
langchain-core
langchain-community
pypdf
pymupdf
sentence-transformers
chromadb
faiss-cpu
langchain-text-splitters
langchain-groq
python-dotenv
fastapi 
uvicorn 
python-multipart 
jinja2
mss 
pytesseract
pillow